+++ CHANGELOG.rst
@@ -1,9 +1,8 @@
-Version 0.1.0 - Development
+Version 0.1.0 - 2001-Aug-24
 =======================================
 
-* First official alpha release.
+* First (very) alpha release.
 * Working for python 2.6 and up, including python 3.
 * Five different applications: HTTP server, RPC server, distributed task queue,
   asynchronous test suite and asynchronous shell.
-* 4 examples spanning all applications.
 * **35 regression tests**